backend/requirements.txt
fastapi
uvicorn
python-multipart
simple-lama-inpainting
pillow
numpy
rembg
opencv-python-headless
